requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0

# Federated Learning
fate-client>=1.11.0
fate-arch>=1.11.0

# LoRA and PEFT
peft>=0.4.0
accelerate>=0.20.0

# Data Processing
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Configuration and Logging
pyyaml>=6.0
tqdm>=4.65.0

# Evaluation Metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2
sacrebleu>=2.3.0

# Optional: For advanced token alignment
editdistance>=0.6.0

# Development Dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipywidgets>=8.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
